Separate base admin list component for reset metadata and display applications
